validate Score value is number

+++ db/migrate/20170727063800_create_scores.rb
@@ -1,7 +1,7 @@
 class CreateScores < ActiveRecord::Migration[5.1]
   def change
     create_table :scores do |t|
-      t.integer :value, :null => true
+      t.float :value, :null => true
       t.references :assignment, foreign_key: true
       t.references :subject, foreign_key: true
